+++ src/lib.rs
@@ -26,8 +26,6 @@ pub mod profile;
 pub mod run;
 mod sys;
 pub mod value;
-#[cfg(feature = "terminal_image")]
-mod viuer;
 
 use std::sync::Arc;